Структура данных для направлений вынесена в отдельный файл
